Shorts changed to byte and clean up to use *.read*(), still in debug mode

+++ Client/src/com/redpois0n/packets/PacketDRIVES.java
@@ -14,9 +14,9 @@ public class PacketDRIVES extends AbstractPacket {
 		for (int i = 0; i < len; i++) {
 			Drive drive = new Drive();
 			drive.setName(slave.readLine());
-			drive.setTotalSpace((short) slave.readInt());
-			drive.setFreeSpace((short) slave.readInt());
-			drive.setUsableSpace((short) slave.readInt());
+			drive.setTotalSpace(slave.readShort());
+			drive.setFreeSpace(slave.readShort());
+			drive.setUsableSpace(slave.readShort());
 			slave.getDrives()[i] = drive;
 		}
 	}
